Добавлен Spring Security

+++ src/main/java/ru/vote/topjava/service/RestaurantService.java
@@ -1,10 +1,13 @@
 package ru.vote.topjava.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 import ru.vote.topjava.model.Restaurant;
 import ru.vote.topjava.model.User;
 import ru.vote.topjava.repository.RestaurantRepository;
+import ru.vote.topjava.repository.UserRepository;
+import ru.vote.topjava.util.SecurityUtil;
 
 import java.util.List;
 
@@ -18,30 +21,39 @@ public class RestaurantService {
         this.restaurantRepository = restaurantRepository;
     }
 
-    // Достать все рестораны
+    // Достать все рестораны (Доступно всем)
     public List<Restaurant> getAllRestaurants(Integer id) {
         return (id == null) ? restaurantRepository.findAll() : restaurantRepository.findAllByIdOwnerRest(id);
     }
 
-    // Достать ресторан по id
+    // Достать ресторан по id (Доступно всем)
     public Restaurant getRestaurantById(int id){
         return restaurantRepository.findById(id).get();
     }
 
-    // Удалить ресторан по id
-    public void delete(int id){
+    // Удалить ресторан по id (Доступно только автору)
+    public void delete(int id, long authUserId){
+        Restaurant restaurant = restaurantRepository.findById(id).get();
+        if (restaurant.getIdOwnerRest() != authUserId){
+            throw new BadCredentialsException("You cannot edit restaurant because restaurant does not exist or you do not have access...");
+        }
         restaurantRepository.deleteById(id);
     }
 
-    // Получить заготовку ресторана (использовать Spring Security)
-    public Restaurant getNewRest(){
-        User admin = new User();
-        Restaurant restaurant = new Restaurant(1,"name", "address", admin);
-        return restaurant;
+    // Получить заготовку ресторана (Доступно только администраторам)
+    public Restaurant getNewRest(long authUserId){
+        if (SecurityUtil.isAdmin()) {
+            Restaurant restaurant = new Restaurant((int) authUserId, "name rest", "address", SecurityUtil.getUser());
+            return restaurant;
+        }
+        throw new BadCredentialsException("You cannot edit restaurant because restaurant does not exist or you do not have access...");
     }
 
-    // Сохранить или обновить ресторан
-    public Restaurant createOrUpdate(Restaurant restaurant){
-        return restaurantRepository.save(restaurant);
+    // Сохранить или обновить ресторан (Доступно только автору)
+    public Restaurant createOrUpdate(Restaurant restaurant, long authUserId){
+        if (restaurant.getIdOwnerRest() == authUserId) {
+            return restaurantRepository.save(restaurant);
+        }
+        throw new BadCredentialsException("You cannot edit restaurant because restaurant does not exist or you do not have access...");
     }
 }
